feat: Add Prisma database client
This commit adds the Prisma database client to the project. It imports the `PrismaClient` from the `@prisma/client` package and declares a global variable `prisma` to hold an instance of the client. If the `NODE_ENV` is not set to "production", the global `prisma` variable is assigned the `prismadb` instance.